00-bootcamp-project/pyproject.toml
[tool.poetry]
name = "00-bootcamp-project"
version = "0.1.0"
description = ""
authors = ["Kan Ouivirach <zkan.cs@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pysftp = "^0.2.9"
psycopg2-binary = "^2.9.9"
google-cloud-bigquery = "^3.34.0"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"